+++ CHANGELOG.md
@@ -6,6 +6,8 @@
  - Fix blast and superblast ("destructible" spell property changed)
  - Add "SuperRing" spell, allow different ring/disc orientations.
  - Add "mana_display" config property, which can be used to display mana as a number rather than a bar.
+ - An empty wand will fill on activate if fill_wands is set to true.
+ - Allow "wand" on Essentials sign for an empty wand.
 
 ## 2.9.3